# server/openapi.yaml
...
openapi: 3.0.3
info:
  title: GWP API
  version: 0.1.0
  description: >
    GWP — Canvas Online as a self-hosted app. Phase 0 ships first-run
    setup, session auth and health probes; Phase 1 adds the organization
    surface (company profile, departments, teams — spec §3/§4). Canvas,
    coaching and AI endpoints arrive in later phases.
  contact:
    name: GWP operators
servers:
  - url: /
    description: Same origin as the served web app.
...
paths:
...
  /api/v1/settings/ai:
    get:
      tags: [settings]
      summary: Read the AI (BYOK) settings — public shape only
      description: >
        Owner/admin only (spec §4 — unlike branding this read is
        privileged). Returns safe metadata only: whether a key is
        configured, the allowlisted endpoint, model and bounds. The API
        key and its encrypted envelope are NEVER returned.
      operationId: getAiSettings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current AI settings (or unconfigured defaults).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSettings"
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Caller is neither owner nor admin.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags: [settings]
      summary: Configure the local AI endpoint and BYOK key
      description: >
        Owner/admin only. `baseUrl` must resolve to a host:port pair in
        the operator's AI_ALLOWED_HOSTS deployment allowlist — admins
        cannot aim the AI client at arbitrary hosts. `apiKey` is stored
        AES-256-GCM encrypted under APP_KEY and never returned; omitting
        it preserves the stored key (explicit clear is DELETE
        /settings/ai/key). Rate-limited per client IP.
      operationId: putAiSettings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiSettingsPut"
      responses:
        "200":
          description: Settings saved — public shape only.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSettings"
        "400":
          description: >
            Body failed validation (strict schema), or baseUrl is not an
            allowlisted destination / carries userinfo / uses http without
            the operator AI_ALLOW_HTTP opt-in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Caller is neither owner nor admin.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Too many settings writes — retry after the bound.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/settings/ai/key:
    delete:
      tags: [settings]
      summary: Explicitly clear the stored AI key
      description: >
        Owner/admin only. Removes the stored credential while keeping the
        endpoint/model/enabled configuration — the separate command makes
        key removal a deliberate act, never a side effect of an omitted
        PUT field.
      operationId: clearAiKey
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Key cleared — public shape with configured=false.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSettings"
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Caller is neither owner nor admin.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Too many settings writes — retry after the bound.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/settings/ai/test:
    post:
      tags: [settings]
      summary: Probe the configured local AI endpoint
      description: >
        Owner/admin only. Sends a fixed synthetic prompt ("Reply with the
        word OK.", 16 output tokens) — never customer data — and reports
        whether the upstream answered, whether it streams SSE, and the
        latency. Failures surface as stable AI_* error codes
        (AI_NOT_CONFIGURED, AI_AUTH_FAILED, AI_BUSY, AI_UNAVAILABLE,
        AI_TIMEOUT, AI_BAD_RESPONSE, AI_REDIRECT_DENIED); upstream error
        bodies are never echoed. Rate-limited per client IP.
      operationId: testAiConnection
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The endpoint answered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiTestResult"
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Caller is neither owner nor admin.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Too many probes — retry after the bound.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "499":
          description: Client cancelled the probe.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: >
            Endpoint reachable but failed: AI_AUTH_FAILED (bad key),
            AI_BAD_RESPONSE (malformed/unsupported answer),
            AI_REDIRECT_DENIED.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: AI_NOT_CONFIGURED / AI_BUSY / AI_UNAVAILABLE.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "504":
          description: AI_TIMEOUT — upstream exceeded the configured bound.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/ai/status:
    get:
      tags: [ai]
      summary: Is local AI usable? (any signed-in user)
      description: >
        Lets every AI surface say "not configured" / "disabled" BEFORE the
        user consents to a run (spec §7.1). Returns exactly two booleans —
        no endpoint, model, key version or key material; the full settings
        document stays owner/admin only (GET /settings/ai). `enabled` is
        true only when a key is configured AND the configuration is on.
      operationId: getAiStatus
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Availability flags.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiStatus"
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/ai/runs:
    post:
      tags: [ai]
      summary: Start a consented AI run on a canvas or coaching session
      description: >
        Canvas assistants (renderer, coach) require `canvasId` and write
        access to the target canvas (spec §4 "Chạy AI trên canvas — theo
        quyền sửa"; foreign/denied canvases answer the uniform 404,
        archived 409). The `oracle` assistant instead requires
        `sessionId` + `transcript` and current report-write authority on
        the coaching session — owner, or the session coach who still
        manages the coachee (spec §6). Requires
        `consent: true` — the caller has seen the internal-AI notice and
        the server re-enforces it. `idempotencyKey` is REQUIRED: a
        retried POST with the same key and same input replays the same
        run (200, replayed:true) instead of launching a second upstream
        call; the same key with different input is 409
        AI_IDEMPOTENCY_CONFLICT. At most 2 active (queued|running) runs
        per company — a third is 429 AI_BUSY. The run row stores metadata
        only (status, input hash, provenance, captured canvas base);
        `notes`, `transcript`, and canvas content are never persisted —
        the transcript exists only inside the request/run and is destroyed
        when inference ends. Progress is streamed on
        GET /ai/runs/{runId}/events; the finished proposal is staged into
        a 15-minute in-memory preview, never auto-applied — an ORACLE
        preview becomes a report only through POST /reports.
      operationId: startAiRun
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiRunStart"
      responses:
        "201":
          description: Run admitted and queued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiRunStartResult"
        "200":
          description: Idempotent replay — same key, same input.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiRunStartResult"
        "400":
          description: INVALID_INPUT (bad fields, missing key) or
            AI_CONSENT_REQUIRED (consent absent/false).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Related actor who lost report-write authority on
            the session (oracle only — unrelated actors get 404).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Canvas or session missing, foreign, or not
            writable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: AI_IDEMPOTENCY_CONFLICT, AI_DISABLED, or
            CANVAS_ARCHIVED.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: AI_BUSY — the per-company active-run cap (2) is
            reached; retry after a run finishes. Also the per-IP rate
            bound.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: AI_NOT_CONFIGURED — no endpoint/key saved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/ai/runs/{runId}:
    get:
      tags: [ai]
      summary: Read run status (creator only)
      description: >
        Returns the run's metadata DTO — status, timestamps, error code,
        usage counters when the endpoint reported them. Only the actor
        who created the run may read it; everyone else gets the uniform
        404 (existence is not revealed).
      operationId: getAiRun
      security:
        - bearerAuth: []
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Run metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiRun"
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Run missing or owned by another actor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/ai/runs/{runId}/events:
    get:
      tags: [ai]
      summary: Stream run progress (SSE, creator only)
      description: >
        `text/event-stream` of {type:"status"|"delta"|"done"|"error"}
        events. The current status is sent first so a late subscriber
        sees where the run is; the stream ends after a terminal event.
        A client disconnect CANCELS a still-active run (spec §7.3 —
        upstream work stops when the caller is gone); terminal runs are
        unaffected.
      operationId: streamAiRunEvents
      security:
        - bearerAuth: []
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: SSE stream; each `data:` frame is a JSON
            AiRunEvent.
          content:
            text/event-stream:
              schema:
                type: string
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Run missing or owned by another actor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/ai/runs/{runId}/preview:
    get:
      tags: [ai]
      summary: Read the staged AI proposal (creator only, 15-min TTL)
      description: >
        The validated proposal held in process memory for 15 minutes
        after the run succeeds — never persisted, never in a log. The
        response carries the `base` (published version + draft revision)
        captured at admission so apply can refuse a moved target. 410
        PREVIEW_EXPIRED covers expired, missing, and never-finished
        alike — the caller must run again.
      operationId: getAiRunPreview
      security:
        - bearerAuth: []
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The staged proposal plus its captured base.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiRunPreview"
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Run missing or owned by another actor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "410":
          description: PREVIEW_EXPIRED — expired or absent.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/ai/runs/{runId}/cancel:
    post:
      tags: [ai]
      summary: Cancel a queued/running run (creator only)
      description: >
        Queued runs become cancelled immediately; a running run's abort
        signal propagates to the upstream call, which stops when the
        endpoint honors it. Terminal runs answer 409 AI_RUN_FINISHED —
        cancellation is not retroactive.
      operationId: cancelAiRun
      security:
        - bearerAuth: []
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Run cancelled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiRun"
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Run missing or owned by another actor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: AI_RUN_FINISHED — run already terminal.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/ai/runs/{runId}/apply:
    post:
      tags: [ai]
      summary: Apply the staged proposal to the draft (creator only, one-shot)
      description: >
        Applies the SERVER-HELD preview — the request carries no canvas
        body, so generated content can never be injected or forged. The
        preview must have no error-severity issues (409
        AI_PROPOSAL_INVALID), every warning must be listed in
        `acceptedWarnings` by issue id (400 AI_WARNINGS_UNACCEPTED), and
        the captured base must still match: a draft that moved since the
        run is 409 AI_BASE_CHANGED, a new published version is 409
        VERSION_CONFLICT. Apply consumes the preview — a second call is
        410 PREVIEW_EXPIRED. The draft is stamped source="ai" with the
        run id; publish remains a separate human step.
      operationId: applyAiRun
      security:
        - bearerAuth: []
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiRunApplyRequest"
      responses:
        "200":
          description: Proposal applied to the draft (draft DTO).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CanvasDraft"
        "400":
          description: >
            INVALID_INPUT (malformed body — including any client-sent
            canvas body) or AI_WARNINGS_UNACCEPTED.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Run missing, owned by another actor, or canvas access lost.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: >
            AI_RUN_NOT_READY (run not succeeded), AI_PROPOSAL_INVALID
            (error-severity issues), AI_BASE_CHANGED (draft moved), or
            VERSION_CONFLICT (published base changed).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "410":
          description: PREVIEW_EXPIRED — expired, consumed, or absent.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          description: Stable machine code (e.g. INVALID_SESSION).
        message:
          type: string
          description: Human-readable, Vietnamese UI copy.
        details:
          description: Optional structured detail (field names, pending list).
        request_id:
          type: string
          format: uuid
          description: Echo of the X-Request-Id response header.
...
    AiStatus:
      type: object
      required: [configured, enabled]
      additionalProperties: false
      properties:
        configured:
          type: boolean
          description: An endpoint key is stored for this company.
        enabled:
          type: boolean
          description: Configured AND switched on — runs can start.
    AiSettings:
      type: object
      required:
        [
          configured,
          enabled,
          baseUrl,
          model,
          timeoutSeconds,
          maxOutputTokens,
          keyVersion,
          updatedAt,
        ]
      additionalProperties: false
      description: >
        Public AI settings shape — safe metadata only. The stored API key
        and its AES-256-GCM envelope are never serialized: `configured`
        is the only signal the credential exists.
      properties:
        configured:
          type: boolean
          description: Whether an API key is stored (server-side).
        enabled:
          type: boolean
        baseUrl:
          type: string
          nullable: true
          description: Allowlisted AI endpoint base URL.
        model:
          type: string
          nullable: true
        timeoutSeconds:
          type: integer
          minimum: 5
          maximum: 180
        maxOutputTokens:
          type: integer
          minimum: 1
          maximum: 8192
        keyVersion:
          type: string
          nullable: true
          description: APP_KEY ring version that sealed the stored key.
        updatedAt:
          type: string
          nullable: true
          format: date-time
    AiSettingsPut:
      type: object
      required: [enabled, baseUrl, model]
      additionalProperties: false
      description: >
        Writable AI settings. `apiKey` is optional — omitting it
        preserves the stored key; clearing is the separate DELETE
        /settings/ai/key command. `null` is rejected for every field.
      properties:
        enabled:
          type: boolean
        baseUrl:
          type: string
          minLength: 1
          maxLength: 500
          description: >
            Must be an allowlisted AI_ALLOWED_HOSTS host:port — no
            userinfo, no query/fragment; http:// needs AI_ALLOW_HTTP.
        model:
          type: string
          minLength: 1
          maxLength: 200
        apiKey:
          type: string
          minLength: 1
          maxLength: 512
          description: Write-only; encrypted at rest, never returned.
        timeoutSeconds:
          type: integer
          minimum: 5
          maximum: 180
          default: 180
        maxOutputTokens:
          type: integer
          minimum: 1
          maximum: 8192
          default: 8192
    AiTestResult:
      type: object
      required: [ok, streaming, model, latencyMs]
      additionalProperties: false
      description: >
        Result of the synthetic connection probe — never carries the
        prompt, the response text, or any credential.
      properties:
        ok:
          type: boolean
          enum: [true]
        streaming:
          type: boolean
          description: Whether the endpoint answered with an SSE stream.
        model:
          type: string
        latencyMs:
          type: integer
          minimum: 0
...